GetChars in non-windows code.

+++ source/common/filesystem/source/files_decompress.cpp
@@ -552,7 +552,7 @@ public:
 			return false;
 		}
 
-		File->Seek(-sizeof(header), FileReader::SeekCur);
+		File->Seek(-(ptrdiff_t)sizeof(header), FileReader::SeekCur);
 
 		FillBuffer();
